import description - meta files
