+++ src/main/java/com/odysseusinc/arachne/datanode/service/AnalysisInfoBuilder.java
@@ -1,0 +1,9 @@
+package com.odysseusinc.arachne.datanode.service;
+
+import com.odysseusinc.arachne.datanode.dto.atlas.CohortDefinition;
+
+public interface AnalysisInfoBuilder {
+
+
+    String generateAnalysisDescription(CohortDefinition definition) ;
+}
