+++ src/main/java/com/odysseusinc/arachne/datanode/service/messaging/LegacyCohortRequestHandler.java
@@ -29,13 +29,13 @@ import com.odysseusinc.arachne.commons.utils.CommonFileUtils;
 import com.odysseusinc.arachne.datanode.dto.atlas.CohortDefinition;
 import com.odysseusinc.arachne.datanode.model.atlas.Atlas;
 import com.odysseusinc.arachne.datanode.model.atlas.CommonEntity;
+import com.odysseusinc.arachne.datanode.service.AnalysisInfoBuilder;
 import com.odysseusinc.arachne.datanode.service.AtlasRequestHandler;
 import com.odysseusinc.arachne.datanode.service.AtlasService;
 import com.odysseusinc.arachne.datanode.service.CommonEntityService;
 import com.odysseusinc.arachne.datanode.service.SqlRenderService;
 import com.odysseusinc.arachne.datanode.service.client.atlas.AtlasClient;
 import com.odysseusinc.arachne.datanode.service.client.portal.CentralSystemClient;
-import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -51,7 +51,7 @@ import java.util.Objects;
 import java.util.stream.Collectors;
 
 import static com.odysseusinc.arachne.commons.api.v1.dto.CommonAnalysisType.COHORT;
-import static com.odysseusinc.arachne.commons.utils.CommonFileUtils.ARACHNE_META_FILE_PREFIX;
+import static com.odysseusinc.arachne.commons.utils.CommonFileUtils.ARACHNE_META_FILE_DESCRIPTION;
 
 
 @Service
@@ -69,6 +69,8 @@ public class LegacyCohortRequestHandler implements AtlasRequestHandler<CommonCoh
             .put("\"", "-quote-")
             .build();
     private static final Logger log = LoggerFactory.getLogger(LegacyCohortRequestHandler.class);
+
+    private final AnalysisInfoBuilder analysisInfoBuilder;
     private final CentralSystemClient centralClient;
     private final GenericConversionService conversionService;
     private final CommonEntityService commonEntityService;
@@ -76,12 +78,9 @@ public class LegacyCohortRequestHandler implements AtlasRequestHandler<CommonCoh
     private final AtlasService atlasService;
 
     @Autowired
-    public LegacyCohortRequestHandler(CentralSystemClient centralClient,
-                                      GenericConversionService conversionService,
-                                      CommonEntityService commonEntityService,
-                                      SqlRenderService sqlRenderService,
-                                      AtlasService atlasService) {
+    public LegacyCohortRequestHandler(AnalysisInfoBuilder analysisInfoBuilder, CentralSystemClient centralClient, GenericConversionService conversionService, CommonEntityService commonEntityService, SqlRenderService sqlRenderService, AtlasService atlasService) {
 
+        this.analysisInfoBuilder = analysisInfoBuilder;
         this.centralClient = centralClient;
         this.conversionService = conversionService;
         this.commonEntityService = commonEntityService;
@@ -117,36 +116,19 @@ public class LegacyCohortRequestHandler implements AtlasRequestHandler<CommonCoh
 
         if (Objects.nonNull(definition)) {
             String content = sqlRenderService.renderSql(definition);
-            String description = generateDescriptionMeta(definition);
+            String description = analysisInfoBuilder.generateAnalysisDescription(definition);
             if (Objects.nonNull(content)) {
                 final String definitionName = definition.getName().trim();
                 final String filteredDefinitionName = filterFileName(definitionName);
                 MockMultipartFile[] files = new MockMultipartFile[]{
                         new MockMultipartFile("file", filteredDefinitionName + CommonFileUtils.OHDSI_JSON_EXT, MediaType.APPLICATION_JSON_VALUE, definition.getExpression().getBytes()),
                         new MockMultipartFile("file", filteredDefinitionName + CommonFileUtils.OHDSI_SQL_EXT, MediaType.TEXT_PLAIN_VALUE, content.getBytes()),
-                        new MockMultipartFile("file", ARACHNE_META_FILE_PREFIX+".description", MediaType.TEXT_PLAIN_VALUE, description.getBytes())
+                        new MockMultipartFile("file", ARACHNE_META_FILE_DESCRIPTION, MediaType.TEXT_PLAIN_VALUE, description.getBytes())
                 };
                 return files;
             }
         }
         return null;
-    }
-
-    private String generateDescriptionMeta(CohortDefinition definition) {
-
-        StringBuilder description = new StringBuilder();
-        if (StringUtils.isNotBlank(definition.getDescription())) {
-            appendLine(description, "Cohort: " + definition.getName());
-            appendLine(description, "Description: " + definition.getDescription());
-        }
-        appendLine(description, "Created by: " + definition.getCreatedBy());
-        appendLine(description, "Creation date: " + definition.getCreatedDate());
-        return description.toString();
-    }
-
-    private static void appendLine(StringBuilder builder, String text){
-        builder.append(text);
-        builder.append(System.lineSeparator());
     }
 
     @Override
